pyproject.toml
[project]
name = "music-poster"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "google-api-python-client>=2.163.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gunicorn>=25.1.0",
    "numpy>=1.24.1",
    "pillow>=12.1.1",
    "python-dotenv>=1.0.0",
    "redis>=5.2.1",
    "requests>=2.32.5",
    "scipy>=1.11.2",
    "spotipy>=2.25.1",
]
